Add async dangling indices support (#50642)
Adds support for writing out dangling indices in an asynchronous way. Also provides an option to
avoid writing out dangling indices at all.

Relates #48701

+++ server/src/main/java/org/elasticsearch/gateway/GatewayMetaState.java
@@ -109,7 +109,11 @@ public class GatewayMetaState implements Closeable {
                             .build());
                     lucenePersistedState = new LucenePersistedState(
                         persistenceWriter, currentTerm, clusterState);
-                    metaStateService.deleteAll(); // delete legacy files
+                    if (DiscoveryNode.isDataNode(settings)) {
+                        metaStateService.unreferenceAll(); // unreference legacy files (only keep them for dangling indices functionality)
+                    } else {
+                        metaStateService.deleteAll(); // delete legacy files
+                    }
                     success = true;
                 } finally {
                     if (success == false) {
